+++ src/main/java/com/depromeet/watni/config/LocalSecurityConfig.java
@@ -23,6 +23,7 @@ public class LocalSecurityConfig extends WebSecurityConfigurerAdapter {
 
     @Override
     protected void configure(HttpSecurity http) throws Exception {
+        http.headers().frameOptions().disable();
         http.csrf().disable();
     }
