--- src/main/java/com/depromeet/watni/domain/accession/AccessionRepository.java
@@ -1,7 +1,0 @@
-package com.depromeet.watni.domain.accession;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccessionRepository extends JpaRepository<Accession, Long> {
-
-}
