+++ src/main/java/com/depromeet/watni/domain/accession/domain/Accession.java
@@ -5,7 +5,7 @@ import com.depromeet.watni.domain.accession.constant.AccessionStatus;
 import com.depromeet.watni.domain.accession.constant.AccessionType;
 import com.depromeet.watni.domain.accession.dto.AccessionResponseDto;
 import com.depromeet.watni.domain.group.domain.Group;
-import com.depromeet.watni.domain.member.Member;
+import com.depromeet.watni.domain.member.domain.Member;
 import lombok.*;
 import javax.persistence.*;
