+++ src/main/java/com/depromeet/watni/domain/accession/dto/AccessionResponseDto.java
@@ -3,7 +3,7 @@ package com.depromeet.watni.domain.accession.dto;
 import com.depromeet.watni.domain.accession.constant.AccessionStatus;
 import com.depromeet.watni.domain.accession.constant.AccessionType;
 import com.depromeet.watni.domain.group.domain.Group;
-import com.depromeet.watni.domain.member.Member;
+import com.depromeet.watni.domain.member.domain.Member;
 import lombok.*;
 
 @Getter
