--- src/main/java/com/depromeet/watni/domain/group/GroupRepository.java
@@ -1,9 +1,0 @@
-package com.depromeet.watni.domain.group;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GroupRepository extends JpaRepository<Group, Long> {
-
-}
