+++ src/main/java/com/depromeet/watni/domain/group/dto/GroupResponseDto.java
@@ -1,20 +1,10 @@
 package com.depromeet.watni.domain.group.dto;
 
+import com.depromeet.watni.domain.conference.domain.Conference;
+import lombok.*;
+
 import java.util.ArrayList;
 import java.util.List;
-
-import javax.persistence.Column;
-import javax.persistence.OneToMany;
-import javax.validation.constraints.NotNull;
-
-import com.depromeet.watni.domain.conference.Conference;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
 @Setter
