+++ src/main/java/com/depromeet/watni/domain/group/service/GroupService.java
@@ -1,8 +1,10 @@
 package com.depromeet.watni.domain.group.service;
 
 import com.depromeet.watni.domain.group.domain.Group;
+import com.depromeet.watni.domain.group.dto.GroupDto;
 import com.depromeet.watni.domain.group.repository.GroupRepository;
 import com.depromeet.watni.exception.BadRequestException;
+import com.depromeet.watni.exception.NotFoundException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -14,5 +16,14 @@ public class GroupService {
 
     public Group selectGroupByGroupId(long groupId) {
         return groupRepository.findById(groupId).orElseThrow(() -> new BadRequestException("NOT FOUND GROUP"));
+    }
+
+    public Group createGroup(GroupDto groupDto) {
+        Group group = Group.builder().name(groupDto.getGroupName()).build();
+        return groupRepository.save(group);
+    }
+
+    public Group getGroup(Long groupId) {
+        return groupRepository.findById(groupId).orElseThrow(() -> new NotFoundException("NOT FOUND GROUP"));
     }
 }
