+++ src/main/java/com/depromeet/watni/domain/groupcode/GroupCode.java
@@ -1,19 +1,9 @@
 package com.depromeet.watni.domain.groupcode;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.OneToOne;
+import com.depromeet.watni.domain.group.domain.Group;
+import lombok.*;
 
-import com.depromeet.watni.domain.group.Group;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import javax.persistence.*;
 
 @Entity
 @Builder
