+++ src/main/java/com/depromeet/watni/domain/manager/dto/ManagerMapper.java
@@ -8,13 +8,13 @@ import org.mapstruct.Mapping;
 @Mapper
 public interface ManagerMapper {
 
-    @Mapping(target = "managerId", source = "managerId")
+    @Mapping(target = "managerId", source = "manager.managerId")
     @Mapping(target = "email", source = "manager.member.email")
     @Mapping(target = "username", source = "manager.member.name")
     @Mapping(target = "group", source = "group")
     MangerResponseDto convert(Manager manager, Group group);
 
-    @Mapping(target = "managerId", source = "managerId")
+    @Mapping(target = "managerId", source = "manager.managerId")
     @Mapping(target = "email", source = "manager.member.email")
     @Mapping(target = "username", source = "manager.member.name")
     MangerResponseDto convert(Manager manager);
