+++ src/main/java/com/depromeet/watni/domain/member/MemberDetail.java
@@ -24,7 +24,7 @@ public class MemberDetail implements UserDetails {
     // TODO append filed;
 
     public MemberDetail(Member member) {
-        this.memberId = member.getId();
+        this.memberId = member.getMemberId();
         this.email = member.getEmail();
         this.name = member.getName();
         this.password = member.getPassword();
