+++ src/main/java/com/depromeet/watni/domain/member/domain/Member.java
@@ -1,10 +1,14 @@
 package com.depromeet.watni.domain.member.domain;
 
 import com.depromeet.watni.base.BaseEntity;
+import com.depromeet.watni.domain.manager.domain.Manager;
 import com.depromeet.watni.domain.member.dto.MemberRequestDto;
+import com.depromeet.watni.domain.participationUser.domain.ParticipationUser;
 import lombok.*;
 
 import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
 
 @Table(name = "member")
 @Entity
@@ -27,6 +31,12 @@ public class Member extends BaseEntity {
 
     @Column(name = "password", nullable = false)
     private String password;
+
+    @OneToMany(mappedBy = "member")
+    private List<Manager> managers = new ArrayList<>();
+
+    @OneToMany(mappedBy = "member")
+    private List<ParticipationUser> participationUsers = new ArrayList<>();
 
     public static Member of(MemberRequestDto memberRequestDto) {
         return Member
